requirements.txt
openai>=1.0.0
pytest>=7.0.0
hypothesis>=6.0.0
python-dotenv>=1.0.0
